Move common models to Common app
