+++ backend/users/backend/administrator/models.py
@@ -5,16 +5,6 @@ from django.utils.translation import gettext_lazy as _
 from base.models import BaseAbstractUser, BaseContentType, BasePermission, BaseGroup
 
 
-class Country(models.Model):
-    id = models.IntegerField(primary_key=True)
-    name = models.CharField(max_length=50)
-
-    class Meta:
-        db_table = "country"
-        verbose_name = _("country")
-        verbose_name_plural = _("countries")
-
-
 class AdminContentType(BaseContentType):
     class Meta(BaseContentType.Meta):
         db_table = "admin_content_type"
