+++ backend/users/backend/common/apps.py
@@ -1,0 +1,6 @@
+from base.apps import BaseAppConfig
+
+
+class CommonAppConfig(BaseAppConfig):
+    name = "common"
+    verbose_name = "Common"
